update ppl.compute for consistency with evaluate implementation

+++ evaluate_perplexity.py
@@ -178,7 +178,7 @@ class Perplexity(evaluate.Metric):
             with torch.no_grad():
                 out_logits = model(encoded_batch, attention_mask=attn_mask).logits
 
-            shift_logits = out_logits[..., :-1, :].contiguous()
+            shift_logits = out_logits[..., :-1, :].contiguous().float()
             shift_labels = labels[..., 1:].contiguous()
             shift_attention_mask_batch = attn_mask[..., 1:].contiguous()
